requirements.txt
# requirements.txt
pyyaml>=6.0
python-dotenv>=1.0
openai>=1.3   # optional—skip if you won't run semantic tests